+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Fixed issues
 
+- [1054](https://github.com/MiKTeX/miktex/issues/1054): Open Terminal window does nothing on elementary OS
 - [1055](https://github.com/MiKTeX/miktex/issues/1055): GFtoPK crashes in supre mode for fonts in certain optical sizes
 
 ## 22.1 - 2022-01-31